Simple use cases of malloc/free
Initial commit of a simple use case that does a malloc and a free.
The basic PointerMapper class is implemented, with the logic for
extracting the id from a fake pointer.
A map of pointers to buffers is maintained in the PointerMapper class.
The generation of the unique id is just an incremental static value.
This value is used as a key to the map.