Only set C++11 for SDK sample code
Previously, the __build_sycl() function set the language version to
C++11, which meant that users who had set a higher version (say 14 or
17) would run into problems. Now, we only set it for the sample code,
instead of all the code. Similarly, I've removed the -Wall that was
sneakily added in. It's still set for the samples, just not for other
projects that happen to use FindComputeCpp.cmake.