Update integration-header on source changes
The device compiler generation of the stub file was not properly
triggered when there were changes associated with files other than
the source file.
This caused the device compiler to not be triggered when related files,
such as included files, were modified, therefore using an incorrect
integration header.
This patch fixes the DEPENDS clause of the add_custom_command so that
it points to the actual source file inependently of the working
directory, and also forces Cmake to trigger the dependency scanning for
the source file.