Add Travis support to SDK (#23)
This commit adds support for the Travis CI system to the SDK. It will download
both ComputeCpp and TriSYCL to the Travis machine before building and testing.
It includes a Travis specific workaround for various maths functions because
Travis exports macros with the same name as the functions, causing problems in
device code.

It uses the Khronos ICD Loader to simulate an OpenCL installation. The tests
then run on the SYCL host device.