Update failing Travis config
The Travis build now uses the OpenCL headers and ICD loader from the
Ubuntu repository, as this is more reliable than the raw Git variants.
Since we don't require an actual OpenCL device, this is still fine.

Additionally, TriSYCL never worked properly for the SDK anyway, since
we don't have proper CMake support for it. Therefore, it's been removed.
This allows us to remove some dependencies from the Travis build script.
Lastly, the greaaterequal problem was fixed in the last ComputeCpp
release.