Virtual Address Space implementation (#33)
This commit implements the virtual pointer interface for buffers, that
enables allocation and  deallocation of pointer-like structures that
are not dereferenciable on the host but can be used to retrieve buffers
and offsets later on.

The PointerMapper is a normal object, unlike the singleton implemented
for the legacy pointer interface. See the README for details.