Rework includes for FindComputeCpp.cmake
Previously, FindComputeCpp.cmake did not properly include the OpenCL
headers for all targets and did not attempt to include the ComputeCpp
headers either. Now, both are added when using the "add_sycl_to_target"
function defined in FindComputeCpp.cmake. The SDK has been updated to be
compliant with this new behaviour.