Update travis config for 0.3.2 (#56)
Hopefully, this config will not break when a new version of ComputeCpp is
released, as it now specifies the version of ComputeCpp to download.

The clang configuration has been removed, as it is failing. Similarly,
the example_vptr test will not be built until it is fixed (in a separate
changeset).