Enable C++ 14 by default (#55)
This change moves the device compiler flag to the start of the flags list,
so that if a user manually provides a different standard through
COMPUTECPP_USER_FLAGS, for example, they will be honoured (as they appear
last). Since more C++ 14 code has turned up in the samples, that is now
the default value for the language version. To keep the CMake output tidy,
the integration header target was renamed _ih.