Add include-after CMake property
This CMake change allows the user to specify that their integration
headers should appear before after the main code of the source file.
This allows users to put enums in the template arguments of their
kernels, even though this is not stricly legal in SYCL code (but lots
of existing code does it).

This is not on by default, because it requires that all kernels used
*must* have a forward definition visible to the host compiler inside
the integration header. For example, you can't have a class local to
a function, or more specifically something like:
    cgh.parallel_for<class kernel_name>(range, [=]() { ... });

This is an advanced property and is unlikely to be required by most
developers, but is there for the special cases.