Reformat with clang-3.9
The clang-format style file has been updated to work with clang-3.9,
and the repository has been formatted accordingly. This required some
changes to be made to the code to fit the formatting as well as some
changes to the style, but this appears to have kept the formatting as
similar as possible while still conforming to the output from
clang-format. No pre-commit hooks have yet been added.