Reorganise SDK contents
This change moves aound many of the SDK files, with several goals in
mind:
* Make it easier to use code in external projects
* Reflect a slightly updated naming convention
* Make it easier to build and test everything in the SDK

These goals are satisfied by this commit. In addition, various smaller
improvements have been made, like some fixes to names that were
incorrect inside source files, some CMake changes and formatting
updates. The changeset is large, but the most important configuration
change is the presence of the new "COMPUTECPP_SDK_BUILD_TESTS", which,
when true, will cause build files for the tests (found in tests/vptr and
tests/legacy-pointer) to be produced.