Change auto to float in Gaussian blur
Doubles are not core in SYCL, so usage of them can fail. Unfortunately
a driver we support has less than stable behaviour when encountering
them on hardware that doesn't support it, choosing to segfault. This
change forces the variable pi to be a float, as auto will deduce the
return value of atan() to be double otherwise. Roll on constexpr maths,
I say!