Update SDK to ComputeCpp v0.5.1 (#85)
More 1.2.1 changes have been integrated, though we're not 100% there
yet. Old code will continue to work but will show deprecation messages
(which I don't think are useful to see when building the SDK).

The opportunity was taken to update the Travis configuration such
that the download script now places ComputeCpp into a known location.
From now on, only the download script should require updates when a
new version of ComputeCpp is released.