Prefix any relative path to make it absolute
If `add_sycl_to_target` is called with relative paths to the source files
then cmake will not automatically adjust these to be absolute paths, and
therefore the device compiler is unable to find the source files. To get
around this, add a step to prefix the relative paths with the currrent
directory and so convert the relative path into an absolute path, which
will be found by the compiler.