Update SDK for next version of ComputeCpp
Samplers now have an enum specifying if they are normalised or not.
Much clearer than a bool.

A larger change is that buffer::get_access() - the function to call
when getting a host accessor - doesn't need the template parameter
access::target::host_buffer, because it is unambiguous (the function
doesn't take a handler, unlike all other get_accesses). This is a
breaking change that has affected a lot of code.

Additionally, a formatting pass was applied over the repository,
which has inflated the size of the diff somewhat.