Update SDK for next version of ComputeCpp
Samplers now have an enum specifying if they are normalised or not.
Much clearer than a bool.

A larger change is that buffer::get_access() - the function to call
when getting a host accessor - doesn't need the template parameter
access::target::host_buffer, because it is unambiguous (the function
doesn't take a handler, unlike all other get_accesses). This is a
breaking change that has affected a lot of code.

The multi-pointer class does not have operator[] in SYCL 1.2.1, so
it has been removed. This was used in the Virtual Pointer class of
the SDK, so users of the vptr will need to follow these changes.
It is simpler now, as we simply get() the OpenCL raw pointer.

Additionally, a formatting pass was applied over the repository,
which has inflated the size of the diff somewhat.