Fix Windows tests (#114)
For more details, see:
https://github.com/codeplaysoftware/computecpp-sdk/issues/112
This is also a necessary (but not sufficient) change in order to make
the tests build on Windows again.

This changes the hardcoded static library paths for Google Test, as
these paths (and file names) differ for Windows builds. CMake is
perfectly able to locate them just by name.

Likewise, instead of linking pthreads directly, the cross-platform CMake
"Threads" are used instead.

Additionally, the C++ language standard for the legacy-pointer tests is
changed from 11 to 14, as Clang will otherwise complain about missing
trailing return types in some MS headers.