Modernise FindComputeCpp.cmake (#118)
This is a very large update to FindComputeCpp.cmake that brings with
it (regrettably) some API changes. See the samples for details, but
the short story is that the add_sycl_to_target function now takes
named parameters, which has helped a lot. Additionally, there have
been bugfixes in the internals of the module, meaning that the include-
after property is more well-behaved.

The other change that is most likely to affect users is the way in which
different bitcode outputs are selected. Previously different options
like "USE_PTX" or similar could be specified, but now an actual value
must be set, and defaults to spir64. This requires a little more work
on the part of the user when a different kind of IR is needed but is
more robust.