Use depfiles in FindComputeCpp.cmake  (#130)
Adds `DEPFILE` to the custom command that invokes the device compiler to
trigger rebuilds of the integration header once an included header changed when
using the Ninja generator and CMake 3.7 and up.

Also, ComputeCpp_ROOT_DIR is set based on the location of the INCLUDE
directory. These locatons are derived from the cache variable passed
by the user, if present, failing that an environment variable, and if
neither are passed, the standard system locations.

Additionally removes a trailing whitespace.