Update depfile generation for device compiler
Ninja has a problem if the name of the object doesn't match what it
expects, so this changes the generation to output a relative path, not
an absolute one. Additionally, the quotes in the optional DEPFILE
variable were causing problems, so they have also been removed.

Many thanks to @jwlawson for finding and fixing this.