Fix sources in multiple targets
Previously, CMake would complain when one source file was used in
multiple targets, as this would cause multiple custom_commands to
have the same output (which appears not to be allowed). To combat
this the target name is now mangled with the source name, forcing
the output to be unique.