Fix handling of $PATH in FindComputeCpp.cmake
It is unlikely that users will want to find compute++ on the path
with ComputeCpp libraries elsewhere in the system. By passing a
parameter to find_program in CMake, we stop it searching the PATH
and it will always instead find the ComputeCpp components in
system paths or the user-provided hint.