Fix the scan example to avoid undefined behavior.
To prevent the original input data from being overwritten, the buffer
constructor with the const pointer input needs to be used. When the
non-const pointer variant is used, the implementation is free to
overwrite the original buffer pointer (if, for example, it chose to use
the CL_MEM_HOST_PTR flag on the underlying OpenCL call), even if it is
being written back to a different data pointer.