Update SDK to ComputeCpp v2.0.0
The name of the ComputeCpp library has changed on Windows. Since
we are supporting MSVC 2019 now, we need to include the compiler
checks to make it work, and also need an update to Google Test
to avoid a deprecation in MSVC.

Additionally, targets which use the ComputeCpp library are now forced
to have their linker language set as C++. The makes sure that (for
example) the platform C++ library is linked to the target, even if
not set by the user code.