Update MSVC compute++ checks
compute++ 2.1.0 is based on Clang 8 and can't parse the Microsoft
C++ standard library headers for MSVC 19.27. This patch makes CMake
try to build a simple program that includes `CL/sycl.hpp`. If the
compilation fails we try to compile with __CUDACC__ and
_HAS_CONDITIONAL_EXPLICIT=0. If compute++ can then build the file,
those flags are added to `COMPUTECPP_DEVICE_COMPILER_FLAGS`.

Co-authored-by: Thales Sabino <thales@codeplay.com>