Allow multiple bitcode targets in COMPUTECPP_BITCODE
Adds support for setting multiple values in COMPUTECPP_BITCODE.
When using ComputeCpp Professional Edition
this allows multiple versions of the same kernel to be available in the same binary.

Display the compiler output when the compiler check fails.
This will give the a better error message to the user.

`get_sycl_target_extension` takes care of setting the filename extension.

Added some documentation.