Update COMPUTECPP_BITCODE to let compiler decide SPIR/SPIR-V by default
Before, COMPUTECPP_BITCODE was set to a default of spir64. As we are aiming for SPIR-V to be the default this need to be changed. COMPUTECPP_BITCODE is now empty by default and when COMPUTECPP_BITCODE is empty, -sycl-target flag will not be used, letting the compiler decide whether to use SPIR or SPIR-V.